move check_extra_op to actions module
